Create core module and move some extensions
