+++ core/src/main/kotlin/com/imashnake/animite/core/extensions/IntExt.kt
@@ -1,3 +1,3 @@
-package com.imashnake.animite.dev.ext
+package com.imashnake.animite.core.extensions
 
 fun Int?.isNullOrZero(): Boolean = this == 0 || this == null
